refactor: rm top segments
